don't unlock file if there is no file...

+++ Tournament.pm
@@ -79,6 +79,7 @@ sub lock {
 sub unlock {
     my ($o) = @_;
     my $file = $o->filename;
+    return unless $file;
     my $lock = "$file.lock";
     unlink $lock;
 }
